+++ sixtyfps_runtime/interpreter/eval.rs
@@ -227,6 +227,10 @@ pub fn eval_expression(e: &Expression, local_context: &mut EvalLocalContext) -> 
                 let x: f64 = eval_expression(&arguments[0], local_context).try_into().unwrap();
                 Value::Number(x.sqrt())
             }
+            Expression::BuiltinFunctionReference(BuiltinFunction::Abs, _) => {
+                let x: f64 = eval_expression(&arguments[0], local_context).try_into().unwrap();
+                Value::Number(x.abs())
+            }
             Expression::BuiltinFunctionReference(BuiltinFunction::Sin, _) => {
                 let x: f64 = eval_expression(&arguments[0], local_context).try_into().unwrap();
                 Value::Number(x.to_radians().sin())